pyproject.toml
[project]
name = "mcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "apify-client>=2.5.0",
    "openai>=2.33.0",
    "pymupdf>=1.27.2.3",
    "python-dotenv>=1.2.2",
    "streamlit>=1.57.0",
]
